+++ src/protocol.class.php
@@ -236,491 +236,3 @@ class ProtocolNode
         return print_r($readableNode, true);
     }
 }
-
-class BinTreeNodeReader
-{
-    private $input;
-    /** @var $key KeyStream */
-    private $key;
-
-    public function resetKey()
-    {
-        $this->key = null;
-    }
-
-    public function setKey($key)
-    {
-        $this->key = $key;
-    }
-
-    public function nextTree($input = null)
-    {
-        if ($input != null) {
-            $this->input = $input;
-        }
-        $firstByte  = $this->peekInt8();
-        $stanzaFlag = ($firstByte & 0xF0) >> 4;
-        $stanzaSize = $this->peekInt16(1) | (($firstByte & 0x0F) << 16);
-        if ($stanzaSize > strlen($this->input)) {
-            throw new Exception("Incomplete message $stanzaSize != " . strlen($this->input));
-        }
-        $this->readInt24();
-        if ($stanzaFlag & 8) {
-            if (isset($this->key)) {
-                $realSize    = $stanzaSize - 4;
-                $this->input = $this->key->DecodeMessage($this->input, $realSize, 0, $realSize);// . $remainingData;
-            } else {
-                throw new Exception("Encountered encrypted message, missing key");
-            }
-        }
-        if ($stanzaSize > 0) {
-            return $this->nextTreeInternal();
-        }
-
-        return null;
-    }
-
-    protected function readNibble() {
-        $byte = $this->readInt8();
-
-        $ignoreLastNibble = (bool) ($byte & 0x80);
-        $size = ($byte & 0x7f);
-        $nrOfNibbles = $size * 2 - (int) $ignoreLastNibble;
-
-        $data = $this->fillArray($size);
-        $string = '';
-
-        for ($i = 0; $i < $nrOfNibbles; $i++) {
-            $byte = $data[(int) floor($i / 2)];
-            $ord = ord($byte);
-
-            $shift = 4 * (1 - $i % 2);
-            $decimal = ($ord & (15 << $shift)) >> $shift;
-
-            switch ($decimal) {
-                case 0:
-                case 1:
-                case 2:
-                case 3:
-                case 4:
-                case 5:
-                case 6:
-                case 7:
-                case 8:
-                case 9:
-                    $string .= $decimal;
-                    break;
-                case 10:
-                case 11:
-                    $string .= chr($decimal - 10 + 45);
-                    break;
-                default:
-                    throw new Exception("Bad nibble: $decimal");
-            }
-        }
-
-        return $string;
-    }
-
-    protected function getToken($token)
-    {
-        $ret     = "";
-        $subdict = false;
-        TokenMap::GetToken($token, $subdict, $ret);
-        if (!$ret) {
-            $token = $this->readInt8();
-            TokenMap::GetToken($token, $subdict, $ret);
-            if (!$ret) {
-                throw new Exception("BinTreeNodeReader->getToken: Invalid token $token");
-            }
-        }
-
-        return $ret;
-    }
-
-    protected function readString($token)
-    {
-        $ret = "";
-
-        if ($token == -1) {
-            throw new Exception("BinTreeNodeReader->readString: Invalid token $token");
-        }
-
-        if (($token > 2) && ($token < 0xf5)) {
-            $ret = $this->getToken($token);
-        } elseif ($token == 0) {
-            $ret = "";
-        } elseif ($token == 0xfc) {
-            $size = $this->readInt8();
-            $ret  = $this->fillArray($size);
-        } elseif ($token == 0xfd) {
-            $size = $this->readInt24();
-            $ret  = $this->fillArray($size);
-        } elseif ($token == 0xfa) {
-            $user   = $this->readString($this->readInt8());
-            $server = $this->readString($this->readInt8());
-            if ((strlen($user) > 0) && (strlen($server) > 0)) {
-                $ret = $user . "@" . $server;
-            } elseif (strlen($server) > 0) {
-                $ret = $server;
-            }
-        } elseif ($token == 0xff) {
-            $ret = $this->readNibble();
-        }
-
-        return $ret;
-    }
-
-    protected function readAttributes($size)
-    {
-        $attributes  = array();
-        $attribCount = ($size - 2 + $size % 2) / 2;
-
-        for ($i = 0; $i < $attribCount; $i++) {
-            $key              = $this->readString($this->readInt8());
-            $value            = $this->readString($this->readInt8());
-            $attributes[$key] = $value;
-        }
-
-        return $attributes;
-    }
-
-    protected function nextTreeInternal()
-    {
-        $token = $this->readInt8();
-        $size  = $this->readListSize($token);
-        $token = $this->readInt8();
-        if ($token == 1) {
-            $attributes = $this->readAttributes($size);
-
-            return new ProtocolNode("start", $attributes, null, "");
-        } elseif ($token == 2) {
-            return null;
-        }
-        $tag        = $this->readString($token);
-        $attributes = $this->readAttributes($size);
-        if (($size % 2) == 1) {
-            return new ProtocolNode($tag, $attributes, null, "");
-        }
-        $token = $this->readInt8();
-        if ($this->isListTag($token)) {
-            return new ProtocolNode($tag, $attributes, $this->readList($token), "");
-        }
-
-        return new ProtocolNode($tag, $attributes, null, $this->readString($token));
-    }
-
-    protected function isListTag($token)
-    {
-        return ($token == 248 || $token == 0 || $token == 249);
-    }
-
-    protected function readList($token)
-    {
-        $size = $this->readListSize($token);
-        $ret  = array();
-        for ($i = 0; $i < $size; $i++) {
-            array_push($ret, $this->nextTreeInternal());
-        }
-
-        return $ret;
-    }
-
-    protected function readListSize($token)
-    {
-        if ($token == 0xf8) {
-            return $this->readInt8();
-        } elseif ($token == 0xf9) {
-            return $this->readInt16();
-        }
-
-        throw new Exception("BinTreeNodeReader->readListSize: Invalid token $token");
-    }
-
-    protected function peekInt24($offset = 0)
-    {
-        $ret = 0;
-        if (strlen($this->input) >= (3 + $offset)) {
-            $ret = ord(substr($this->input, $offset, 1)) << 16;
-            $ret |= ord(substr($this->input, $offset + 1, 1)) << 8;
-            $ret |= ord(substr($this->input, $offset + 2, 1)) << 0;
-        }
-
-        return $ret;
-    }
-
-    protected function readInt24()
-    {
-        $ret = $this->peekInt24();
-        if (strlen($this->input) >= 3) {
-            $this->input = substr($this->input, 3);
-        }
-
-        return $ret;
-    }
-
-    protected function peekInt16($offset = 0)
-    {
-        $ret = 0;
-        if (strlen($this->input) >= (2 + $offset)) {
-            $ret = ord(substr($this->input, $offset, 1)) << 8;
-            $ret |= ord(substr($this->input, $offset + 1, 1)) << 0;
-        }
-
-        return $ret;
-    }
-
-    protected function readInt16()
-    {
-        $ret = $this->peekInt16();
-        if ($ret > 0) {
-            $this->input = substr($this->input, 2);
-        }
-
-        return $ret;
-    }
-
-    protected function peekInt8($offset = 0)
-    {
-        $ret = 0;
-        if (strlen($this->input) >= (1 + $offset)) {
-            $sbstr = substr($this->input, $offset, 1);
-            $ret   = ord($sbstr);
-        }
-
-        return $ret;
-    }
-
-    protected function readInt8()
-    {
-        $ret = $this->peekInt8();
-        if (strlen($this->input) >= 1) {
-            $this->input = substr($this->input, 1);
-        }
-
-        return $ret;
-    }
-
-    protected function fillArray($len)
-    {
-        $ret = "";
-        if (strlen($this->input) >= $len) {
-            $ret         = substr($this->input, 0, $len);
-            $this->input = substr($this->input, $len);
-        }
-
-        return $ret;
-    }
-}
-
-class BinTreeNodeWriter
-{
-    private $output;
-    /** @var $key KeyStream */
-    private $key;
-
-    public function resetKey()
-    {
-        $this->key = null;
-    }
-
-    public function setKey($key)
-    {
-        $this->key = $key;
-    }
-
-    public function StartStream($domain, $resource)
-    {
-        $attributes = array();
-
-        $attributes["to"]       = $domain;
-        $attributes["resource"] = $resource;
-        $this->writeListStart(count($attributes) * 2 + 1);
-
-        $this->output .= "\x01";
-        $this->writeAttributes($attributes);
-
-        return "WA" . $this->writeInt8(1) . $this->writeInt8(5) . $this->flushBuffer();
-    }
-
-    /**
-     * @param ProtocolNode $node
-     * @param bool         $encrypt
-     *
-     * @return string
-     */
-    public function write($node, $encrypt = true)
-    {
-        if ($node == null) {
-            $this->output .= "\x00";
-        } else {
-            $this->writeInternal($node);
-        }
-
-        return $this->flushBuffer($encrypt);
-    }
-
-    /**
-     * @param ProtocolNode $node
-     */
-    protected function writeInternal($node)
-    {
-        $len = 1;
-        if ($node->getAttributes() != null) {
-            $len += count($node->getAttributes()) * 2;
-        }
-        if (count($node->getChildren()) > 0) {
-            $len += 1;
-        }
-        if (strlen($node->getData()) > 0) {
-            $len += 1;
-        }
-        $this->writeListStart($len);
-        $this->writeString($node->getTag());
-        $this->writeAttributes($node->getAttributes());
-        if (strlen($node->getData()) > 0) {
-            $this->writeBytes($node->getData());
-        }
-        if ($node->getChildren()) {
-            $this->writeListStart(count($node->getChildren()));
-            foreach ($node->getChildren() as $child) {
-                $this->writeInternal($child);
-            }
-        }
-    }
-
-    protected function parseInt24($data)
-    {
-        $ret = ord(substr($data, 0, 1)) << 16;
-        $ret |= ord(substr($data, 1, 1)) << 8;
-        $ret |= ord(substr($data, 2, 1)) << 0;
-        return $ret;
-    }
-
-    protected function flushBuffer($encrypt = true)
-    {
-        $size = strlen($this->output);
-        $data = $this->output;
-        if ($this->key != null && $encrypt) {
-            $bsize = $this->getInt24($size);
-            //encrypt
-            $data     = $this->key->EncodeMessage($data, $size, 0, $size);
-            $len      = strlen($data);
-            $bsize[0] = chr((8 << 4) | (($len & 16711680) >> 16));
-            $bsize[1] = chr(($len & 65280) >> 8);
-            $bsize[2] = chr($len & 255);
-            $size     = $this->parseInt24($bsize);
-        }
-        $ret          = $this->writeInt24($size) . $data;
-        $this->output = '';
-        return $ret;
-    }
-
-    protected function getInt24($length)
-    {
-        $ret = '';
-        $ret .= chr((($length & 0xf0000) >> 16));
-        $ret .= chr((($length & 0xff00) >> 8));
-        $ret .= chr(($length & 0xff));
-        return $ret;
-    }
-
-    protected function writeToken($token)
-    {
-        if ($token < 0xf5) {
-            $this->output .= chr($token);
-        } elseif ($token <= 0x1f4) {
-            $this->output .= "\xfe" . chr($token - 0xf5);
-        }
-    }
-
-    protected function writeJid($user, $server)
-    {
-        $this->output .= "\xfa";
-        if (strlen($user) > 0) {
-            $this->writeString($user);
-        } else {
-            $this->writeToken(0);
-        }
-        $this->writeString($server);
-    }
-
-    protected function writeInt8($v)
-    {
-        $ret = chr($v & 0xff);
-
-        return $ret;
-    }
-
-    protected function writeInt16($v)
-    {
-        $ret = chr(($v & 0xff00) >> 8);
-        $ret .= chr(($v & 0x00ff) >> 0);
-
-        return $ret;
-    }
-
-    protected function writeInt24($v)
-    {
-        $ret = chr(($v & 0xff0000) >> 16);
-        $ret .= chr(($v & 0x00ff00) >> 8);
-        $ret .= chr(($v & 0x0000ff) >> 0);
-
-        return $ret;
-    }
-
-    protected function writeBytes($bytes)
-    {
-        $len = strlen($bytes);
-        if ($len >= 0x100) {
-            $this->output .= "\xfd";
-            $this->output .= $this->writeInt24($len);
-        } else {
-            $this->output .= "\xfc";
-            $this->output .= $this->writeInt8($len);
-        }
-        $this->output .= $bytes;
-    }
-
-    protected function writeString($tag)
-    {
-        $intVal  = -1;
-        $subdict = false;
-        if (TokenMap::TryGetToken($tag, $subdict, $intVal)) {
-            if ($subdict) {
-                $this->writeToken(236);
-            }
-            $this->writeToken($intVal);
-            return;
-        }
-        $index = strpos($tag, '@');
-        if ($index) {
-            $server = substr($tag, $index + 1);
-            $user   = substr($tag, 0, $index);
-            $this->writeJid($user, $server);
-        } else {
-            $this->writeBytes($tag);
-        }
-    }
-
-    protected function writeAttributes($attributes)
-    {
-        if ($attributes) {
-            foreach ($attributes as $key => $value) {
-                $this->writeString($key);
-                $this->writeString($value);
-            }
-        }
-    }
-
-    protected function writeListStart($len)
-    {
-        if ($len == 0) {
-            $this->output .= "\x00";
-        } elseif ($len < 256) {
-            $this->output .= "\xf8" . chr($len);
-        } else {
-            $this->output .= "\xf9" . $this->writeInt16($len);
-        }
-    }
-}
